FEATURE: HTTP API Survey Node
This add a new survey node, which we will perform an HTTP API Request.
This will provide two fields on the Survey UI to configure the API:
- API URL: enter the url for the HTTP API (GET Method)
- Timeout: set timeout of the API request

When using API node, the result will be matched against list of
patterns and branches, this will allow to define what to do based on
the API results (HTTP content) and select the appropriate branching.

A list of HTTP variables will be passed out to identify the current call
session, e.g. contact, current variable, see the documentation for more
information.